% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Placeholder function}
\usage{
foo(x)
}
\arguments{
\item{x}{argument}
}
\value{
Prints to console, called for side-effects.
}
\description{
Placeholder function
}
\examples{
foo(5)
foo("a")
}
